added edit item functionality
